+++ AuroraEditor/Base/Documents/Model/WorkspaceDocument/AuroraEditorDocumentController.swift
