+++ AuroraEditor/Base/Documents/Model/WorkspaceDocument/WorkspaceDocument+Search.swift
