+++ AuroraEditor/Base/Documents/Model/WorkspaceDocument/WorkspaceDocument+Selection.swift
