+++ AuroraEditor/Base/Documents/Model/WorkspaceDocument/WorkspaceDocument+Tabs.swift
